chore(rawdb): move plugin/evm/rawdb to plugin/evm/customrawdb
