+++ core/state/pruner/pruner.go
@@ -38,7 +38,7 @@ import (
 	"time"
 
 	"github.com/ava-labs/coreth/core/state/snapshot"
-	customrawdb "github.com/ava-labs/coreth/plugin/evm/rawdb"
+	"github.com/ava-labs/coreth/plugin/evm/customrawdb"
 	"github.com/ava-labs/libevm/common"
 	"github.com/ava-labs/libevm/core/rawdb"
 	"github.com/ava-labs/libevm/core/types"
