+++ plugin/evm/customrawdb/accessors_snapshot_ext.go
@@ -1,7 +1,7 @@
 // (c) 2025, Ava Labs, Inc. All rights reserved.
 // See the file LICENSE for licensing terms.
 
-package rawdb
+package customrawdb
 
 import (
 	"github.com/ava-labs/libevm/common"
