+++ plugin/evm/customrawdb/accessors_state_sync_test.go
@@ -1,7 +1,7 @@
 // (c) 2023, Ava Labs, Inc. All rights reserved.
 // See the file LICENSE for licensing terms.
 
-package rawdb
+package customrawdb
 
 import (
 	"testing"
